@@ -17,13 +17,41 @@ public class JsonConvention : ISetupConvention
     public void Register(IConventionContext context)
     {
         if (RuntimeInformation.IsOSPlatform(OSPlatform.Create("Browser"))) return;
+        var applicationName = context.Get<string>("ApplicationName");
         context.AppendApplicationConfiguration(
-            configurationBuilder => new[] { new ConfigurationBuilderDelegateResult("appsettings.json", LoadJsonFile(configurationBuilder, "appsettings.json")) }
+            configurationBuilder =>
+            {
+                ConfigurationBuilderDelegateResult[] results =
+                [
+                    new ("appsettings.json", LoadJsonFile(configurationBuilder, "appsettings.json")),
+                ];
+
+#if  NET10_0_OR_GREATER
+                return applicationName is {Length: > 0} ? [
+                    ..results,
+                    new ($"{applicationName}.json", LoadJsonFile(configurationBuilder, $"{applicationName}.json")),
+                ] : results;
+#else
+                return results;
+#endif
+            }
         );
         context.AppendEnvironmentConfiguration(
-            (configurationBuilder, environment) => new[]
+            (configurationBuilder, environment) =>
             {
-                new ConfigurationBuilderDelegateResult($"appsettings.{environment}.json", LoadJsonFile(configurationBuilder, $"appsettings.{environment}.json")),
+                ConfigurationBuilderDelegateResult[] results =
+                [
+                    new ($"appsettings.{environment}.json", LoadJsonFile(configurationBuilder, $"appsettings.{environment}.json")),
+                ];
+
+#if  NET10_0_OR_GREATER
+                return applicationName is {Length: > 0} ? [
+                    ..results,
+                    new ($"{applicationName}.{environment}.json", LoadJsonFile(configurationBuilder, $"{applicationName}.{environment}.json")),
+                ] : results;
+#else
+                return results;
+#endif
             }
         );
     }
