@@ -16,24 +16,45 @@ public class YamlConvention : ISetupConvention
     public void Register(IConventionContext context)
     {
         if (RuntimeInformation.IsOSPlatform(OSPlatform.Create("Browser"))) return;
+        var applicationName = context.Get<string>("ApplicationName");
         context.AppendApplicationConfiguration(
             configurationBuilder =>
             {
-                return new ConfigurationBuilderDelegateResult[]
-                {
-                    new("appsettings.yaml", LoadYamlFile(configurationBuilder, "appsettings.yaml")),
-                    new("appsettings.yml", LoadYamlFile(configurationBuilder, "appsettings.yml")),
-                };
+                ConfigurationBuilderDelegateResult[] results =
+                [
+                    new ("appsettings.yaml", LoadYamlFile(configurationBuilder, "appsettings.yaml")),
+                    new ("appsettings.yml", LoadYamlFile(configurationBuilder, "appsettings.yml"))
+                ];
+
+#if  NET10_0_OR_GREATER
+                return applicationName is {Length: > 0} ? [
+                    ..results,
+                    new($"{applicationName}.yaml", LoadYamlFile(configurationBuilder, $"{applicationName}.yaml")),
+                    new($"{applicationName}.yml", LoadYamlFile(configurationBuilder, $"{applicationName}.yml")),
+                ] : results;
+#else
+                return results;
+#endif
             }
         );
         context.AppendEnvironmentConfiguration(
             (configurationBuilder, environment) =>
             {
-                return new ConfigurationBuilderDelegateResult[]
-                {
+                ConfigurationBuilderDelegateResult[] results =
+                [
                     new($"appsettings.{environment}.yaml", LoadYamlFile(configurationBuilder, $"appsettings.{environment}.yaml")),
                     new($"appsettings.{environment}.yml", LoadYamlFile(configurationBuilder, $"appsettings.{environment}.yml")),
-                };
+                ];
+
+#if  NET10_0_OR_GREATER
+                return applicationName is {Length: > 0} ? [
+                    ..results,
+                    new($"{applicationName}.{environment}.yaml", LoadYamlFile(configurationBuilder, $"{applicationName}.{environment}.yaml")),
+                    new($"{applicationName}.{environment}.yml", LoadYamlFile(configurationBuilder, $"{applicationName}.{environment}.yml")),
+                ] : results;
+#else
+                return results;
+#endif
             }
         );
     }
