@@ -47,13 +47,13 @@ public class RocketWebApplicationTests(ITestOutputHelper outputHelper) : AutoFak
                                     .ConfigureRocketSurgery();
         var configuration = (IConfigurationRoot)host.Services.GetRequiredService<IConfiguration>();
 
-        #if NET10_0_OR_GREATER
-        configuration.Providers.OfType<JsonConfigurationProvider>().Count().ShouldBe(8);
-        configuration.Providers.OfType<YamlConfigurationProvider>().Count().ShouldBe(12);
-        #else
+#if NET10_0_OR_GREATER
+        configuration.Providers.OfType<JsonConfigurationProvider>().Count().ShouldBe(12);
+        configuration.Providers.OfType<YamlConfigurationProvider>().Count().ShouldBe(24);
+#else
         configuration.Providers.OfType<JsonConfigurationProvider>().Count().ShouldBe(6);
         configuration.Providers.OfType<YamlConfigurationProvider>().Count().ShouldBe(12);
-        #endif
+#endif
     }
 
     [Fact]
@@ -103,13 +103,13 @@ public class RocketWebApplicationTests(ITestOutputHelper outputHelper) : AutoFak
                                     .ConfigureRocketSurgery(z => z.ExceptConvention(typeof(YamlConvention)));
         var configuration = (IConfigurationRoot)host.Services.GetRequiredService<IConfiguration>();
 
-        #if NET10_0_OR_GREATER
-        configuration.Providers.OfType<JsonConfigurationProvider>().Count().ShouldBe(8);
+#if NET10_0_OR_GREATER
+        configuration.Providers.OfType<JsonConfigurationProvider>().Count().ShouldBe(12);
         configuration.Providers.OfType<YamlConfigurationProvider>().Count().ShouldBe(0);
-        #else
+#else
         configuration.Providers.OfType<JsonConfigurationProvider>().Count().ShouldBe(6);
         configuration.Providers.OfType<YamlConfigurationProvider>().Count().ShouldBe(0);
-        #endif
+#endif
     }
 
     [Fact]
@@ -121,12 +121,12 @@ public class RocketWebApplicationTests(ITestOutputHelper outputHelper) : AutoFak
 
         var configuration = (IConfigurationRoot)host.Services.GetRequiredService<IConfiguration>();
 
-        #if NET10_0_OR_GREATER
+#if NET10_0_OR_GREATER
         configuration.Providers.OfType<JsonConfigurationProvider>().Count().ShouldBe(2);
-        configuration.Providers.OfType<YamlConfigurationProvider>().Count().ShouldBe(12);
-        #else
+        configuration.Providers.OfType<YamlConfigurationProvider>().Count().ShouldBe(24);
+#else
         configuration.Providers.OfType<JsonConfigurationProvider>().Count().ShouldBe(0);
         configuration.Providers.OfType<YamlConfigurationProvider>().Count().ShouldBe(12);
-        #endif
+#endif
     }
 }
